Modify send_ha_discovery.sh to accept files and numbers as parameters.

+++ drivers/send_ha_discovery.sh
@@ -6,22 +6,73 @@
 DRIVER=$(mktemp /tmp/send_ha_discovery.driver.XXXXXX)
 TELEGRAM=$(mktemp /tmp/send_ha_discovery.json.XXXXXX)
 
-cat <<< "$1" > $DRIVER
-cat <<< "$2" > $TELEGRAM
+DEBUG=false
+if [ "$1" = "--debug" ]
+then
+    shift
+    DEBUG=true
+fi
 
-echo // Telegram content /////////////////////////////////////////////
-./build/xmq $TELEGRAM for-each '/telegram/node()' --shell='echo "${..}  ...  ${.}"'
-echo // DRIVER as xmq /////////////////////////////////////////////
-./build/xmq $DRIVER to-xmq --compact
-echo // DRIVER as json /////////////////////////////////////////////
-./build/xmq $DRIVER to-json
-echo // DRIVER as xml /////////////////////////////////////////////
-./build/xmq $DRIVER to-xml
-# Generate a HA meter config. Reads the first telegram but does not acutally use the content.
-echo // HA CONFIG //////////////////////////////////////////
-./build/xmq $DRIVER transform --stringparam=file=$TELEGRAM drivers/to-ha-discovery.xslq to-json
+if [ -z "$1" ] || [ -z "$2" ]
+then
+    echo "Usage: send_ha_discovery.sh {--debug} [driver.xmq] [telegram.json]"
+    echo "       send_ha_discovery.sh {--debug} [driver as string] [telegram as string]"
+    echo "You can also specify 1 to extract test telegram 1 from driver."
+    exit 0
+fi
+
+if [ -f "$1" ]
+then
+    # Copy source file into file.
+    cat "$1" > $DRIVER
+else
+    # Copy parameter string into file.
+    cat <<< "$1" > $DRIVER
+fi
+
+if [ -f "$2" ]
+then
+    # Copy source file into file.
+    cat "$2" > $TELEGRAM
+else
+    # Check if parameters is a number, eg 1 or 2 or 3 etc.
+    if [ "$2" -eq "$2" ]
+    then
+        ARGS=$(xmq drivers/src/iperl.xmq select '(//args)[1]' to-text)
+        HEX=$(xmq drivers/src/iperl.xmq select '(//telegram)[1]' to-text)
+        wmbusmeters --format=json $HEX $ARGS > $TELEGRAM
+    else
+        # Copy parameter string into file.
+        cat <<< "$2" > $TELEGRAM
+    fi
+fi
+
+if [ "$DEBUG" = "true" ]
+then
+    echo "// Telegram content /////////////////////////////////////////////"
+    ./build/xmq $TELEGRAM for-each '/telegram/node()' --shell='echo "${..}  ...  ${.}"'
+
+    echo "// DRIVER as xmq /////////////////////////////////////////////"
+    ./build/xmq $DRIVER to-xmq --compact
+
+    echo "// DRIVER as json /////////////////////////////////////////////"
+    ./build/xmq $DRIVER to-json
+
+    echo "// DRIVER as xml /////////////////////////////////////////////"
+    ./build/xmq $DRIVER to-xml
+fi
+
+# Generate a HA meter config from the driver and the first telegram.
+# The first telegram is needed because drivers can handle many different meters
+# and contain a lot of fields that are not relevant for this meter.
+
+echo "// HA CONFIG //////////////////////////////////////////"
+
+echo ./build/xmq $DRIVER transform --stringparam=file=$TELEGRAM drivers/to-ha-discovery.xslq
+
 echo ///////////////////////////////////////////////////////
 
+cat $TELEGRAM
 
-rm -f $DRIVER_XMQ
-rm -f $TELEGRAM_JSON
+#rm -f $DRIVER
+#rm -f $TELEGRAM
